Require outlierDetection for EDS locality failover (#14099)
* Require outlierDetection for EDS locality failover

Without outlierDetection, failover is not safe as Envoy cannot detect
hosts are unhealthy. Because of this, all traffic will be locked into
the same locality, even if there are no healthy hosts in the locality.
To prevent this, we will only apply failover settings when outlier
detection is present. While the logic to get the outlierDetection is
fairly involved, computationally it is not too expensive (from
performance testing and just looking at the code).

* Add negative test case