Backport changes to new installer (#14210)
* Remove Injector dependency on MeshConfig

The only field it references is one that is not configurable anyways, so
we can hard code it in the template safely. If it does need to be
configured in the future, we should read from values directly rather
than MeshConfig.

Backport of f6f501c10c9ba78ead5645803608e8fbf85294e5 from
istio/installer

* add pdb for sidecar_injector

backport of #138 from istio/installer

* Add neverInjectSelector and alwaysInjectSelector

backport of istio/install #165

* fix(dashboards): ensure irate used for success rate calc

Backport of istio/install #166

* Modify dns refresh rate to 5m

backport of istio/install #152

* Add toleration to installer

backport of istio/installer #170

* Fix goldens