Change the version tag for base image from latest (#13875)
The bookinfo-review sample app uses websphere-liberty:latest to create a base
image. This PR changes the latest tag to reflect an actual version. As a best
practise, we do not want to use the latest tag since that could point to different
versions as the release of the base image changes.

Partially Fixes:#13262