Add readiness probe to proxy to check Envoy status (#14928)
* Add readiness probe to proxy to check Envoy status

Currently we just check that a LDS and CDS response was received and
that there is correct inbound listeners. However, there are a lot of
cases where this is not sufficient, and Envoy is still not ready. For
instance, if the listeners are waiting on RDS or SDS response. In this
case, Envoy will not actually accept traffic as expected.

Envoy exposes its current state in the /server_info endpoint, so we can
read that and ensure we don't mark it as ready until it has completed
initialization.

* Fix lint errors

* Add license