Don't produce duplicate wildcard host matches (#15628)
Currently, if there are two https services, both will attempt to create
this wildcard listener. Later, one of them will be rejected and logged
as an outbound conflict. With this change we check to make sure that not
only does the filter chain we are building not already have a wildcard,
but the existing listener also does not have a wildcard.