Disable frozen config store (#15671)
The frozen config store was meant to detect when we were modifying the
configs returned from the config store. However, this caused issues,
because we actually do want to modify - in particular, we sort the
configs. Freezing + sorting can lead to obscure segmentation faults due
to the freeze library using unsafe pointers, which was causing test
failures.