Fix consul monitoring test flakes (#15821)
Previously, the tests would wait for some period of time and check if it
got any updates. This timing sensitive test fails often when CPU is
throttled or slow, especially in -race or coverage mode. Additionally,
one of the tests depended on the order of a map which is undefined.

This change sorts the map output to be deterministic and changes the
test to poll for success rather than wait and check.