Cleanup unused resources, regenerate (#305)
* Cleanup GOPATH

Restore mount for local (faster testing)

* Add ports to kind, for exposing ingress to local development.

This is WIP, intended for multi-cluster testing and simplified local
dev.

* Remove unused/out-of-scope configs.

* Remove unused configs.

Simplify the steps,  the makefile can be used as an example.

Added a (commented) out curl test for knative and fortio

* Cleanup and align with the decision to use istio-system in 1.3

* KNative test must use a separate namespace.

Everything else reverted to istio-system as decided, but knative has the
requirement to run alongside istio-system.

This adds a kustomization for installing istio-micro namespace and
values for manual kube-inject, using the corresponding namespace.

* Cleanup

- remove old files that are not part of the release plan.
- use 'gen-' prefix for generated files
- remove the micro-bundle, instead use 'minimal' (matching 1.2 profile)
and ingress.
- update all generated files to match current values.

* Switch defaults to istio-system, since this is what we're shipping.

We'll still test with multiple namespaces, but not blocking and not part
presubmit

* Add a build option to support istio 1.2 labels for injection.

Using istio-env and multiple namespaces not in scope for 1.3

* Update the template

* Further cleanup to align with the reduced scope.

Test are using istio-system, one namespace by default.

"make test" runs all the one-namespace tests.

* The 'run-minikube' target to test demo.

The 'integration' test is the slowest - we need the minikube test just
to verify the install and test are working with minikube and VM, so we
don't break support for it for users who test on minikube. No need to
run the largest test.

The exact same test is run in a remote-docker env with kind, so we can
compare run speeds and reliability.

* Flaky test: split crds creation for knative test

* Fix namespace to istio-system.

Only one test left with multiple namespaces, to verify we don't break
for 1.4

* Fix and add canary tests

* Knative ingress must use istio-discovery in same namespace.

WIP to use env for kustomization

* One more 'default to istio-system"

* Remove adding the canary test to unblock, will be added later

* Remove 'until' - it actually loops forever, and timeout doesn't
seem to happen

* Remove duplicated reachability test

* more descriptive name, isolate knative test

* Adjust allocations for knative test, it was not fitting in circle

* And the cpu allocations