Fix unbuffered error channel usage (#16624)
This fixes a common issue where we have an error channel that is
unbuffered. Later we have a select{}, and if that selects another case
the channel will never be read from, causing a deadlock

See https://github.com/istio/istio/pull/15897 for a real life case where
this caused Pilot to die.

Fixes https://github.com/istio/istio/issues/15902