Initial version of istiod. (#17944)
* Initial version of istiod.

Based on the original hyperistio and tests. Includes:
- galley
- pilot
- injector
- cert generation using k8s apiserver

I've tried to keep k8s dependencies isolated, since some components may
be built without k8s. Even if we ship only one binary, it's good to keep
modules separated.

* Add boilerplate

* Add the 'var' files - the istiod version of the configs.

* Remove temp file

* Fix lint on some files.

Add back the SDS (still on UDS)

* SDS server now using the same GRPC server as Pilot.

No use of envoy - instead use the K8S cert and the existing secure grpc
startup of Pilot.

Note: for upgrade scenarios we have 2 options, one is to add a sidecar
with the spiffe certificates, the other is to shift traffic to istiod.
The second is cleaner and we'll use it initially.

* Fix cert, no need to ask for client cert

* Fix review feedback, lint.

* More fixes

* More lint/test fixes.

The unit tests run with an old K8S that can't sign certs. Will be
covered by integration tests.

* One more lint

* Istiod is not currently using flags

* Remove the var dir - will use the defaults for now

* Exclude istiod from test flags (we may end flags based on dicussion)