mixer: remove explicit flush in Opencensus tracing adapter (#18074)
* mixer: remove explicit flush in Opencensus tracing adapter

Currently, the Opencensus tracing adapter manually calls flush with each
batch of tracespan instances that are passed to it. The implementation
of the flush method is such that it calls (*Bundler).Flush, which blocks
until all batches of spans in the bundler are written to the tracing
backend.

Given the flush method is blocking (i.e. only one goroutine can be
flushing at any point in time), this makes the
(*Handler).HandleTraceSpan method blocking, which in turn prevents
concurrent calls to mixer's Report endpoint.

The blocking in the tracing adapter ultimately results in increased
overall latency to the Report endpoint, which can trip the rate limiting
functionality, resulting in unnecessary throttling, even on requests
that do not contain tracing information.

Remove the explicit flushing in the Opencensus tracing adapter, instead
relying on the flushing functionality provided in the underlying tracing
library, which defaults to flushing once every two seconds.

Fixes #18042.

Signed-off-by: Nick Travers <n.e.travers@gmail.com>

* Fix formatting