Add sidecar injector to pilot (#18898)
Part of
https://docs.google.com/document/d/1jO2oM9zdm9h7tUqOpnAG3dh1xavtW1L7zQ5LyraYRAs.

Right now this change is not activated, because the default Pilot
deployment will not have the injector mounted. This will require changes
to istio/installer to set this up. It is intentional that istio/istio
helm installer is not changed here, as we probably want to leave that
untouched.